audio/video: consistent gstreamer handling
This adds some more helper RAII functions to fix the following bug when a video
file cannot be found:

    GLib-GObject-CRITICAL **: g_object_unref: assertion 'G_IS_OBJECT (object)' failed

This also makes some error handling and event handling consistent.

+++ src/detail/video/gstdecoderimpl.h
@@ -50,9 +50,9 @@ public:
 
     virtual bool playing() const;
 
-    virtual bool volume(double volume);
+    virtual bool volume(int volume);
 
-    virtual double volume() const;
+    virtual int volume() const;
 
     std::string error_message() const;
